Time: 13 ms (60.67%), Space: 114.9 MB (57.3%) - LeetHub

+++ 2285-maximum-total-importance-of-roads/2285-maximum-total-importance-of-roads.java
@@ -9,7 +9,7 @@ class Solution {
         Arrays.sort(degree);
         long maxImportance=0;
 
-        int label=1;
+        long label=1;
         for(int i=0;i<n;i++){
             maxImportance += degree[i] * label;
             label++;
